Implemented diagram layout computation.
Geometrical information is extracted from the diagram by calculating the
action of the diagram limits on geometrical coordinates. By padding the
geometrical information with half an identity on all sides, we not only
simplify the code, but automatically get the padding of the diagram to a
rectangular shape later. The positions of the geometrical points of the
diagram in space is determined by a solver based on a cassowary linear
program.