Attachments.
Geometric coordinates can be directly translated into a boundary path,
which then is used for attachment. This procedure will very likely work
for 3D diagrams without modification.

A list of attachment options is generated, but currently only unique
attachments are executed, since the UI for selecting attachment options
is not yet implemented.

This also fixes a bug in the original attachment code: Composing 1-cells
A -> B and B -> A would yield an ill-typed diagram before.