Dramatic performance improvement in layout solver.
In codimension 0, the layout algorithm can simply return the codimension
0 coordinate of the input point, which removes all constraints in that
codimension from consideration.

Average constraints in the highest dimension are the strongest, so we
can employ a union-find datastructure to unify two variables in that
dimension when we would have constrained them to be equal with an
average constraint before.

Together, these optimizations cut down layout for a simple medium sized
2D diagram on my machine from 1.5s to about 10ms. Without this
optimization, the algorithm would likely have been inadequate for any
real life diagrams. This should also improve the performance in case of
higher dimensional diagrams.